Force cat as cocos2d object
The problem was that the dynamic type inference was looking into UIKit as well, which had an object with a "rotation" property of type CGVector - which made it throw the error that incorrect types were being allocated.

This forces the library to always get a CCNode pointer and use the rotation from that class.